add unit tests, simplify legacy
